fix bugs

+++ grails-app/domain/com/ybg/app/mshow/UserAccount.groovy
@@ -10,7 +10,7 @@ class UserAccount {
         usedMoney nullable: true
     }
 
-    Float totalMoney = 0//全部金额
-    Float restMoney = 0//可用金额
-    Float usedMoney = 0//己使用金额
+    Integer totalMoney = 0//全部金额
+    Integer restMoney = 0//可用金额
+    Integer usedMoney = 0//己使用金额
 }
